+++ src/handlers/rss.rs
@@ -11,13 +11,22 @@ use std::sync::Arc;
 
 pub async fn show(Host(host): Host, State(state): State<Arc<AppState>>) -> Response {
     let mut pages: Vec<Page> = Vec::new();
-    for entry in glob("./pages/**/*.md").expect("Failed to read glob pattern") {
+    for entry in
+        glob(&format!("{}/pages/**/*.md", state.path)).expect("Failed to read glob pattern")
+    {
         match entry {
             Ok(path) => {
-                let file_path: String = path.into_os_string().into_string().unwrap();
-                if file_path != "pages/home.md" {
-                    let url: String = file_path.replace("pages/", "").replace(".md", "");
-                    let page: Page = Page::new(url, read_file(&file_path), state.mk_parser_options);
+                let file_name: String = path
+                    .file_name()
+                    .unwrap()
+                    .to_os_string()
+                    .into_string()
+                    .unwrap();
+                if file_name != "home.md" {
+                    let file_content: String =
+                        read_file(&path.into_os_string().into_string().unwrap());
+                    let url: String = file_name.replace(".md", "");
+                    let page: Page = Page::new(url, file_content, state.mk_parser_options);
                     if page.published {
                         pages.push(page);
                     }